[[ Bug 20259 ]] Cast binary field length to a long before calculating encoded length of output buffer

+++ revdb/src/sqlite_connection.cpp
@@ -439,7 +439,7 @@ bool queryCallback(void *p_context, int p_placeholder, DBBuffer& p_output)
 		else
 		{
 			// According to documentation in sqlitedecode.cpp, this is the required size of output buffer
-			t_escaped_string = malloc(2 + (257 * t_parameter_value . length) / 254);
+			t_escaped_string = malloc(2 + (257 * (long)t_parameter_value . length) / 254);
 			t_escaped_string_length = sqlite_encode_binary((const unsigned char *)t_parameter_value . sptr, t_parameter_value . length, (unsigned char *)t_escaped_string);
 		}
 	}
